Renamed GeneralizedLedgerEntry.{h,cpp} to InternalLedgerEntry.{h,cpp} and update includes
