+++ src/ledger/InMemoryLedgerTxnRoot.h
@@ -1,4 +1,4 @@
-#include "ledger/GeneralizedLedgerEntry.h"
+#include "ledger/InternalLedgerEntry.h"
 #include "ledger/LedgerTxn.h"
 #include "ledger/LedgerTxnImpl.h"
 #include "xdr/Stellar-ledger-entries.h"
