+++ src/ledger/InternalLedgerEntry.h
